Use bilinear interpolation to read u,v at point x,y
This avoids a number of corner-cases that come up when using average of cell corner points weighted by distance. It is also significantly simpler and should be faster, although this was not tested.

This results in slightly different stream-lines in complex cases but eyeballing the streamlines.png it seems "equivalent"