Scale wind vector to take one step along either axis
This avoids traces being chopped up when they fail to step out of their current cell so they land on the cell that they've just marked as used. This is particularly apparent when the wind is blowing at a 45° angle to the grid and the vector needs to have length sqrt(2) to be sure to enter the next cell.