Round x,y instead of floor before testing used pixel
This detects cases where two traces are just on either side of the grid indices and round to the same value but floor to different values. The result is a more uniform stream field when there are many long traces such as for high-altitude winds.

Also removed redundant test in each trace-iteration

Also fixed indent in .eslintrc to allow the constant condition in `while(true)`